+++ src/main/kotlin/klib/exceptions/KonfParseException.kt
@@ -1,0 +1,11 @@
+package klib.exceptions
+
+/**
+ * Custom exception
+ *
+ * @param reason
+ *
+ * @since 0.2.2
+ * @author Thomas Obernosterer
+ */
+class KonfParseException(reason: String) : Exception("KonfParseException: Cannot parse konf; Reason: $reason")